mapper: use define_method to define each_* methods for a has_n relation
Should have used `define_method` from the start, but at that time I was under
the impression that `define_method` could not be used to define methods that
take a block, so I went with `class_eval`.  `class_eval`, however, was causing
tests to fail on 1.9.2 (#4) for some weird reason, so we switch to
`define_method`.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>